requirement.txt
# Core Data
pandas
numpy
pyarrow
fastparquet

# ML - Model 1
scikit-learn
xgboost
torch
torchvision

# Experiment Tracking
mlflow

# Explainability
shap

# Visualization / EDA
matplotlib
seaborn
plotly

# API Serving
fastapi
uvicorn

# Notebook
jupyter
ipykernel

streamlit
python-dotenv
openai
huggingface_hub
datasets
